# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"tonari-backend","version":"0.0.1"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/ingest/syosetu":{"post":{"tags":["ingest"],"summary":"Ingest Syosetu","operationId":"ingest_syosetu_ingest_syosetu_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestSyosetuRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chapter-translations/":{"post":{"tags":["chapter-translations"],"summary":"Create Chapter Translation","operationId":"create_chapter_translation_chapter_translations__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterTranslationCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterTranslationOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chapter-translations/{ct_id}":{"get":{"tags":["chapter-translations"],"summary":"Get Chapter Translation","operationId":"get_chapter_translation_chapter_translations__ct_id__get","parameters":[{"name":"ct_id","in":"path","required":true,"schema":{"type":"integer","title":"Ct Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterTranslationOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chapter-translations/{ct_id}/segments":{"get":{"tags":["chapter-translations"],"summary":"List Translation Segments","operationId":"list_translation_segments_chapter_translations__ct_id__segments_get","parameters":[{"name":"ct_id","in":"path","required":true,"schema":{"type":"integer","title":"Ct Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TranslationSegmentOut"},"title":"Response List Translation Segments Chapter Translations  Ct Id  Segments Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/prompts/":{"get":{"tags":["prompts"],"summary":"List Prompts","description":"List all prompts globally with optional search.","operationId":"list_prompts_prompts__get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Q"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedPromptsOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["prompts"],"summary":"Create Prompt","description":"Create a new global prompt.","operationId":"create_prompt_prompts__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/prompts/{prompt_id}":{"get":{"tags":["prompts"],"summary":"Get Prompt","description":"Get a prompt and its metadata including latest version.","operationId":"get_prompt_prompts__prompt_id__get","parameters":[{"name":"prompt_id","in":"path","required":true,"schema":{"type":"integer","title":"Prompt Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptDetailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["prompts"],"summary":"Update Prompt","description":"Update prompt metadata (name and/or description).","operationId":"update_prompt_prompts__prompt_id__patch","parameters":[{"name":"prompt_id","in":"path","required":true,"schema":{"type":"integer","title":"Prompt Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptUpdateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/prompts/{prompt_id}/versions":{"get":{"tags":["prompts"],"summary":"List Prompt Versions","description":"List all versions of a prompt.","operationId":"list_prompt_versions_prompts__prompt_id__versions_get","parameters":[{"name":"prompt_id","in":"path","required":true,"schema":{"type":"integer","title":"Prompt Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedPromptVersionsOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["prompts"],"summary":"Append Prompt Version","description":"Append a new version to a prompt.","operationId":"append_prompt_version_prompts__prompt_id__versions_post","parameters":[{"name":"prompt_id","in":"path","required":true,"schema":{"type":"integer","title":"Prompt Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptVersionCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptVersionOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/prompts/{prompt_id}/versions/{version_id}":{"get":{"tags":["prompts"],"summary":"Get Prompt Version","description":"Get a specific version of a prompt.","operationId":"get_prompt_version_prompts__prompt_id__versions__version_id__get","parameters":[{"name":"prompt_id","in":"path","required":true,"schema":{"type":"integer","title":"Prompt Id"}},{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptVersionOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/prompts/works/{work_id}/prompt":{"get":{"tags":["prompts"],"summary":"Get Work Prompt","description":"Get the prompt assigned to a work.","operationId":"get_work_prompt_prompts_works__work_id__prompt_get","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptDetailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/":{"get":{"tags":["works"],"summary":"Search Works","operationId":"search_works_works__get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Q"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedWorksOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/import":{"post":{"tags":["works"],"summary":"Import Work","operationId":"import_work_works_import_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkImportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/{work_id}":{"get":{"tags":["works"],"summary":"Get Work","operationId":"get_work_works__work_id__get","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/{work_id}/chapters":{"get":{"tags":["works"],"summary":"List Chapters For Work","operationId":"list_chapters_for_work_works__work_id__chapters_get","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedChaptersOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/{work_id}/chapters/{chapter_id}":{"get":{"tags":["works"],"summary":"Get Chapter For Work","operationId":"get_chapter_for_work_works__work_id__chapters__chapter_id__get","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}},{"name":"chapter_id","in":"path","required":true,"schema":{"type":"integer","title":"Chapter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterDetailOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/{work_id}/scrape-chapters":{"post":{"tags":["works"],"summary":"Request Chapter Scrape","operationId":"request_chapter_scrape_works__work_id__scrape_chapters_post","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterScrapeRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterScrapeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/{work_id}/chapters/{chapter_id}/translation":{"get":{"tags":["works"],"summary":"Get Chapter Translation State","operationId":"get_chapter_translation_state_works__work_id__chapters__chapter_id__translation_get","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}},{"name":"chapter_id","in":"path","required":true,"schema":{"type":"integer","title":"Chapter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterTranslationStateOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["works"],"summary":"Reset Chapter Translation","operationId":"reset_chapter_translation_works__work_id__chapters__chapter_id__translation_delete","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}},{"name":"chapter_id","in":"path","required":true,"schema":{"type":"integer","title":"Chapter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterTranslationStateOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/{work_id}/chapters/{chapter_id}/regenerate-segments":{"post":{"tags":["works"],"summary":"Regenerate Chapter Segments","description":"Regenerate all segments for a chapter, discarding existing translations.\n\nThis endpoint deletes all existing segments for all translations of the chapter\nand recreates them based on the current chapter text. This is useful when the\nchapter source text changes and needs to be re-segmented.","operationId":"regenerate_chapter_segments_works__work_id__chapters__chapter_id__regenerate_segments_post","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}},{"name":"chapter_id","in":"path","required":true,"schema":{"type":"integer","title":"Chapter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/{work_id}/chapters/{chapter_id}/translate/stream":{"get":{"tags":["works"],"summary":"Stream Chapter Translation","operationId":"stream_chapter_translation_works__work_id__chapters__chapter_id__translate_stream_get","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}},{"name":"chapter_id","in":"path","required":true,"schema":{"type":"integer","title":"Chapter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChapterDetailOut":{"properties":{"id":{"type":"integer","title":"Id"},"work_id":{"type":"integer","title":"Work Id"},"idx":{"type":"integer","title":"Idx"},"sort_key":{"type":"number","title":"Sort Key"},"title":{"type":"string","title":"Title"},"normalized_text":{"type":"string","title":"Normalized Text"}},"type":"object","required":["id","work_id","idx","sort_key","title","normalized_text"],"title":"ChapterDetailOut"},"ChapterOut":{"properties":{"id":{"type":"integer","title":"Id"},"work_id":{"type":"integer","title":"Work Id"},"idx":{"type":"integer","title":"Idx"},"sort_key":{"type":"number","title":"Sort Key"},"title":{"type":"string","title":"Title"}},"type":"object","required":["id","work_id","idx","sort_key","title"],"title":"ChapterOut"},"ChapterScrapeErrorItem":{"properties":{"chapter":{"type":"number","title":"Chapter"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["chapter","reason"],"title":"ChapterScrapeErrorItem"},"ChapterScrapeRequest":{"properties":{"start":{"type":"number","minimum":1.0,"title":"Start","description":"Starting chapter number (supports decimals)"},"end":{"type":"number","minimum":1.0,"title":"End","description":"Ending chapter number (supports decimals)"},"force":{"type":"boolean","title":"Force","description":"If true, rescrape existing chapters","default":false}},"type":"object","required":["start","end"],"title":"ChapterScrapeRequest"},"ChapterScrapeResponse":{"properties":{"work_id":{"type":"integer","title":"Work Id"},"start":{"type":"number","title":"Start"},"end":{"type":"number","title":"End"},"force":{"type":"boolean","title":"Force"},"status":{"type":"string","title":"Status"},"requested":{"type":"integer","title":"Requested"},"created":{"type":"integer","title":"Created"},"updated":{"type":"integer","title":"Updated"},"skipped":{"type":"integer","title":"Skipped"},"errors":{"items":{"$ref":"#/components/schemas/ChapterScrapeErrorItem"},"type":"array","title":"Errors"}},"type":"object","required":["work_id","start","end","force","status","requested","created","updated","skipped"],"title":"ChapterScrapeResponse"},"ChapterTranslationCreate":{"properties":{"chapter_id":{"type":"integer","title":"Chapter Id"},"cache_policy":{"type":"string","title":"Cache Policy","default":"reuse"},"params":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Params"}},"type":"object","required":["chapter_id"],"title":"ChapterTranslationCreate"},"ChapterTranslationOut":{"properties":{"id":{"type":"integer","title":"Id"},"chapter_id":{"type":"integer","title":"Chapter Id"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","chapter_id","status"],"title":"ChapterTranslationOut"},"ChapterTranslationStateOut":{"properties":{"chapter_translation_id":{"type":"integer","title":"Chapter Translation Id"},"status":{"type":"string","title":"Status"},"segments":{"items":{"$ref":"#/components/schemas/TranslationSegmentOut"},"type":"array","title":"Segments"}},"type":"object","required":["chapter_translation_id","status","segments"],"title":"ChapterTranslationStateOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IngestSyosetuRequest":{"properties":{"novel_id":{"type":"string","pattern":"^n[a-z0-9]+$","title":"Novel Id","description":"Syosetu novel identifier (e.g. n4811fg)"},"chapter":{"type":"integer","minimum":1.0,"title":"Chapter","description":"Chapter number (1-indexed) to fetch"}},"type":"object","required":["novel_id","chapter"],"title":"IngestSyosetuRequest"},"PaginatedChaptersOut":{"properties":{"items":{"items":{"$ref":"#/components/schemas/ChapterOut"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"limit":{"type":"integer","title":"Limit"},"offset":{"type":"integer","title":"Offset"}},"type":"object","required":["items","total","limit","offset"],"title":"PaginatedChaptersOut"},"PaginatedPromptVersionsOut":{"properties":{"items":{"items":{"$ref":"#/components/schemas/PromptVersionOut"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"limit":{"type":"integer","title":"Limit"},"offset":{"type":"integer","title":"Offset"}},"type":"object","required":["items","total","limit","offset"],"title":"PaginatedPromptVersionsOut"},"PaginatedPromptsOut":{"properties":{"items":{"items":{"$ref":"#/components/schemas/PromptOut"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"limit":{"type":"integer","title":"Limit"},"offset":{"type":"integer","title":"Offset"}},"type":"object","required":["items","total","limit","offset"],"title":"PaginatedPromptsOut"},"PaginatedWorksOut":{"properties":{"items":{"items":{"$ref":"#/components/schemas/WorkOut"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"limit":{"type":"integer","title":"Limit"},"offset":{"type":"integer","title":"Offset"}},"type":"object","required":["items","total","limit","offset"],"title":"PaginatedWorksOut"},"PromptCreateRequest":{"properties":{"name":{"type":"string","minLength":1,"title":"Name","description":"Prompt name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Optional description"}},"type":"object","required":["name"],"title":"PromptCreateRequest"},"PromptDetailOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"owner_work_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Owner Work Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"latest_version":{"anyOf":[{"$ref":"#/components/schemas/PromptVersionOut"},{"type":"null"}]}},"type":"object","required":["id","name","created_at","updated_at"],"title":"PromptDetailOut","description":"Prompt with latest version info included."},"PromptOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"owner_work_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Owner Work Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","name","created_at","updated_at"],"title":"PromptOut"},"PromptUpdateRequest":{"properties":{"name":{"anyOf":[{"type":"string","minLength":1},{"type":"null"}],"title":"Name","description":"New prompt name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"New description"}},"type":"object","title":"PromptUpdateRequest"},"PromptVersionCreateRequest":{"properties":{"model":{"type":"string","minLength":1,"title":"Model","description":"Model name (e.g., gpt-4)"},"template":{"type":"string","minLength":1,"title":"Template","description":"F-string template for the prompt"},"parameters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Parameters","description":"Optional metadata parameters"},"created_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created By","description":"Optional creator identifier"}},"type":"object","required":["model","template"],"title":"PromptVersionCreateRequest"},"PromptVersionOut":{"properties":{"id":{"type":"integer","title":"Id"},"prompt_id":{"type":"integer","title":"Prompt Id"},"version_number":{"type":"integer","title":"Version Number"},"model":{"type":"string","title":"Model"},"template":{"type":"string","title":"Template"},"parameters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Parameters"},"created_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created By"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","prompt_id","version_number","model","template","created_at"],"title":"PromptVersionOut"},"TranslationSegmentOut":{"properties":{"id":{"type":"integer","title":"Id"},"start":{"type":"integer","title":"Start"},"end":{"type":"integer","title":"End"},"order_index":{"type":"integer","title":"Order Index"},"src":{"type":"string","title":"Src"},"tgt":{"type":"string","title":"Tgt"},"flags":{"items":{"type":"string"},"type":"array","title":"Flags"}},"type":"object","required":["id","start","end","order_index","src","tgt"],"title":"TranslationSegmentOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WorkImportRequest":{"properties":{"url":{"type":"string","minLength":1,"format":"uri","title":"Url","description":"URL to the work homepage or chapter"},"force":{"type":"boolean","title":"Force","description":"Force re-sync even if the work exists","default":false}},"type":"object","required":["url"],"title":"WorkImportRequest"},"WorkOut":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"},"source_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Id"},"source_meta":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Source Meta"}},"type":"object","required":["id","title"],"title":"WorkOut"}}}}